add-abstract-class
